fix: polish all 9 v1.1 visualisation charts to Stephen Few standard
- streaming_live_feed: proper HH:MM timestamps, non-overlapping direct labels
- online_regime_evolution: use #4E79A7 (blue) for Calm instead of green,
  staggered end-of-line labels to prevent overlap
- online_regime_transitions: direct labels placed at longest regime run,
  use blue for Calm regime colour
- online_detection_lag: use primary blue for batch line
- alert_timeline: replace cluttered triangle markers with vertical lines
  at alpha 0.45 for cleaner crash-zone visualisation
- alert_threshold_monitor: fix label precision (0.025 not 0.03)
- All charts: Stephen Few muted palette, horizontal gridlines only,
  no box/frame, direct labels, bbox_inches=tight, dpi=150

+++ examples/online_regime_charts.py
@@ -50,7 +50,7 @@ def chart_regime_evolution() -> None:
     """Regime posterior probabilities evolving over time."""
     rng = np.random.default_rng(42)
     true_regimes, returns = _simulate_regime_data(300)
-    timestamps = np.arange(len(returns))
+    timestamps = np.arange(len(returns), dtype=float)
 
     # Simulate posterior probabilities (smoothed one-hot with noise)
     probs = np.zeros((len(returns), 3))
